PurchaseAnalysis/requirements.txt
Django==1.11
django-extensions==2.2.9
djangorestframework==3.10.3
ipdb==0.13.2
ipython==7.14.0
psycopg2==2.8.5